ploop-copy.c: u32 passed as u64 to send_image_block() #PSBM-47575
Error in send_image_block (ploop-copy.c:481): Error from pread() size=18446744071667974144 pos=41589669888: Bad address

+++ lib/ploop-copy.c
@@ -747,7 +747,7 @@ int ploop_copy_next_iteration(struct ploop_copy_handle *h,
 
 		for (pos = e.start; pos < e.end; ) {
 			ssize_t n;
-			int copy = e.end - pos;
+			__u64 copy = e.end - pos;
 
 			if (copy > h->cluster)
 				copy = h->cluster;
